fix(update): use 'long qint64' and not just 'int' for assets size
fix(update): use 'long qint64' and not just 'int' for assets size

+++ src/backend/DownloadManager.h
@@ -57,7 +57,7 @@ class DownloadManager: public QObject
     Q_OBJECT
 public:
     explicit DownloadManager(QObject *parent = nullptr);
-    void append(const QUrl &url, const QString &filename, const int filesize);
+    void append(const QUrl &url, const QString &filename, const qint64 filesize);
     void append(const QUrl &url, const QString &filename);
     void append(const QStringList &urls);
     static QString saveFileName(const QUrl &url);
@@ -89,7 +89,7 @@ private:
     QNetworkAccessManager manager;
     QQueue<QUrl> downloadQueue;
     QQueue<QString> filenameQueue;
-    QQueue<int> filesizeQueue;
+    QQueue<qint64> filesizeQueue;
     QNetworkReply *currentDownload = nullptr;
     QFile output;
     QElapsedTimer downloadTimer;
